chore: move package config to internal/config
