+++ api/exec_test.go
@@ -17,10 +17,11 @@ import (
 	"github.com/gorilla/websocket"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/tsuru/rpaas-operator/config"
+	"golang.org/x/net/http2"
+
+	"github.com/tsuru/rpaas-operator/internal/config"
 	"github.com/tsuru/rpaas-operator/internal/pkg/rpaas"
 	"github.com/tsuru/rpaas-operator/internal/pkg/rpaas/fake"
-	"golang.org/x/net/http2"
 )
 
 var h2cClient = &http.Client{
